erp-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.32
pydantic==2.9.1

Pillow==10.4.0

python-multipart==0.0.9
